api/db.py
import os
from supabase import create_client, Client
from dotenv import load_dotenv

load_dotenv()
_client = None


def get_client() -> Client:
    global _client
    if _client is None:
        _client = create_client(
            os.environ["SUPABASE_URL"],
            os.environ["SUPABASE_KEY"]
        )
    return _client
