api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.9.1
httpx==0.27.2
websockets==13.1
python-dotenv==1.0.0
pydantic==2.5.0
